chore: clean up transfer
